# services/@tables/service/spec/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Tables Service API",
    "description": "Dynamic data tables API with Yjs CRDT real-time collaboration",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/tables/{tableId}/rows": {
      "get": {
        "operationId": "listRows",
        "tags": [
          "rows"
        ],
        "description": "List rows in a table (paginated)",
        "parameters": [
          {
            "schema": {
              "type": "number"
            },
            "in": "query",
            "name": "offset",
            "required": false,
            "description": "Pagination offset"
          },
          {
            "schema": {
              "type": "number"
            },
            "in": "query",
            "name": "limit",
            "required": false,
            "description": "Page size (default 50)"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "sortBy",
            "required": false,
            "description": "Column ID to sort by"
          },
          {
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "enum": [
                    "asc"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "desc"
                  ]
                }
              ]
            },
            "in": "query",
            "name": "sortOrder",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "tableId",
            "required": true,
            "description": "Table ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "rows",
                    "total",
                    "offset",
                    "limit"
                  ],
                  "properties": {
                    "rows": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "id",
                          "data",
                          "createdAt",
                          "updatedAt"
                        ],
                        "properties": {
                          "id": {
                            "type": "string",
                            "description": "Row ID (UUID)"
                          },
                          "data": {
                            "type": "object",
                            "description": "Column values keyed by column ID",
                            "additionalProperties": {}
                          },
                          "createdAt": {
                            "type": "string",
                            "description": "Creation timestamp (ISO 8601)"
                          },
                          "updatedAt": {
                            "type": "string",
                            "description": "Last update timestamp (ISO 8601)"
                          }
                        }
                      }
                    },
                    "total": {
                      "type": "number",
                      "description": "Total number of rows"
                    },
                    "offset": {
                      "type": "number",
                      "description": "Current offset"
                    },
                    "limit": {
                      "type": "number",
                      "description": "Page size"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "insertRow",
        "tags": [
          "rows"
        ],
        "description": "Insert a new row",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "data": {
                    "type": "object",
                    "description": "Initial column values",
                    "additionalProperties": {}
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "tableId",
            "required": true,
            "description": "Table ID"
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "data",
                    "createdAt",
                    "updatedAt"
                  ],
                  "properties": {
                    "id": {
                      "type": "string",
                      "description": "Row ID (UUID)"
                    },
                    "data": {
                      "type": "object",
                      "description": "Column values keyed by column ID",
                      "additionalProperties": {}
                    },
                    "createdAt": {
                      "type": "string",
                      "description": "Creation timestamp (ISO 8601)"
                    },
                    "updatedAt": {
                      "type": "string",
                      "description": "Last update timestamp (ISO 8601)"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/tables/{tableId}/rows/{rowId}": {
      "patch": {
        "operationId": "updateRow",
        "tags": [
          "rows"
        ],
        "description": "Update a row (partial update)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "data"
                ],
                "properties": {
                  "data": {
                    "type": "object",
                    "description": "Column values to update",
                    "additionalProperties": {}
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "tableId",
            "required": true,
            "description": "Table ID"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "rowId",
            "required": true,
            "description": "Row ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "data",
                    "createdAt",
                    "updatedAt"
                  ],
                  "properties": {
                    "id": {
                      "type": "string",
                      "description": "Row ID (UUID)"
                    },
                    "data": {
                      "type": "object",
                      "description": "Column values keyed by column ID",
                      "additionalProperties": {}
                    },
                    "createdAt": {
                      "type": "string",
                      "description": "Creation timestamp (ISO 8601)"
                    },
                    "updatedAt": {
                      "type": "string",
                      "description": "Last update timestamp (ISO 8601)"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteRow",
        "tags": [
          "rows"
        ],
        "description": "Delete a row",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "tableId",
            "required": true,
            "description": "Table ID"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "rowId",
            "required": true,
            "description": "Row ID"
          }
        ],
        "responses": {
          "204": {
            "description": "Default Response"
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/tables/{tableId}/rows/bulk": {
      "post": {
        "operationId": "bulkRowOperation",
        "tags": [
          "rows"
        ],
        "description": "Perform bulk operations on rows",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "operation",
                  "rowIds"
                ],
                "properties": {
                  "operation": {
                    "anyOf": [
                      {
                        "type": "string",
                        "enum": [
                          "delete"
                        ]
                      }
                    ]
                  },
                  "rowIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "Row IDs to operate on"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "tableId",
            "required": true,
            "description": "Table ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "affected"
                  ],
                  "properties": {
                    "affected": {
                      "type": "number",
                      "description": "Number of rows affected"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3005",
      "description": "Development server"
    }
  ]
}